tryout backport
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@49613 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-3.18/include/net/inet_hashtables.h
@@ -386,7 +386,7 @@ static inline struct sock *__inet_lookup_skb(struct inet_hashinfo *hashinfo,
 }
 
 int __inet_hash_connect(struct inet_timewait_death_row *death_row,
-			struct sock *sk, u32 port_offset,
+			struct sock *sk, u64 port_offset,
 			int (*check_established)(struct inet_timewait_death_row *,
 						 struct sock *, __u16,
 						 struct inet_timewait_sock **),
